🚧 Provide PointValue logic
- Experiment with Different PointValue types.
- Set DataPointType on DataPoint instead on value. It reduces the unnecessary abstractions.
- PointTypes will be useful in future when we will provide Rendering Options, or we will try to attach Point to specific Widget.

+++ srf-commons/src/main/kotlin/org/reggster/srfcommons/async/PointValue.kt
@@ -3,6 +3,6 @@ package org.reggster.srfcommons.async
 data class PointValue(
     var dsId: Int,
     var dpId: Int,
-    var value: Int,
+    var value: Double,
     var time: Long?
 )
